Support deleting project in restore failed state
